Separates hook and context
