+++ src/context/EditModeContext.tsx
@@ -1,29 +1,8 @@
-import { createContext, useContext, useState } from "react";
-import type { ReactNode } from "react";
+import { createContext } from "react";
 
 type EditModeContextType = {
   isEditMode: boolean;
   toggleEditMode: () => void;
 };
 
-const EditModeContext = createContext<EditModeContextType | undefined>(undefined);
-
-export function EditModeProvider({ children }: { children: ReactNode }) {
-  const [isEditMode, setIsEditMode] = useState(false);
-
-  const toggleEditMode = () => setIsEditMode((prev) => !prev);
-
-  return (
-    <EditModeContext.Provider value={{ isEditMode, toggleEditMode }}>
-      {children}
-    </EditModeContext.Provider>
-  );
-}
-
-export function useEditMode() {
-  const context = useContext(EditModeContext);
-  if (!context) {
-    throw new Error("useEditMode must be used within an EditModeProvider");
-  }
-  return context;
-}
+export const EditModeContext = createContext<EditModeContextType | undefined>(undefined);
