+++ src/context/EditModeProvider.tsx
@@ -1,0 +1,15 @@
+import { useState } from "react";
+import type { ReactNode } from "react";
+import { EditModeContext } from "./EditModeContext";
+
+export function EditModeProvider({ children }: { children: ReactNode }) {
+  const [isEditMode, setIsEditMode] = useState(false);
+
+  const toggleEditMode = () => setIsEditMode((prev) => !prev);
+
+  return (
+    <EditModeContext.Provider value={{ isEditMode, toggleEditMode }}>
+      {children}
+    </EditModeContext.Provider>
+  );
+}
